demo: address review feedback on ty integration
- Use ty Workspace.codeActions (superset of Diagnostic.codeAction) so
  workspace-derived fixes like import suggestions are surfaced too.
- Extract the duplicated LSP range-overlap helper into demo/shared/ranges.ts,
  shared by the Ruff and ty servers.
- build-ty-wasm.sh: hard-reset the ruff cache to the fetched ref instead of
  swallowing pull failures, so the built artifact always matches RUFF_REF.
