+++ demo/python/ruffServer.ts
@@ -7,6 +7,7 @@ import {
     DiagnosticSeverity,
     TextDocumentSyncKind,
 } from "vscode-languageserver-protocol";
+import { rangesOverlap } from "../shared/ranges";
 
 type RuffFix = NonNullable<RuffDiagnostic["fix"]>;
 
@@ -162,15 +163,4 @@ function wholeDocumentEdit(text: string, newText: string): LSP.TextEdit {
         },
         newText,
     };
-}
-
-function rangesOverlap(a: LSP.Range, b: LSP.Range): boolean {
-    return !(isBefore(a.end, b.start) || isBefore(b.end, a.start));
-}
-
-function isBefore(a: LSP.Position, b: LSP.Position): boolean {
-    if (a.line !== b.line) {
-        return a.line < b.line;
-    }
-    return a.character < b.character;
 }
